feat: scaffold media_assistant module and extract shared code
Extract browser, volume, system from agent/tools/ into shared/.
Create media_assistant/ skeleton with all subpackages, config, and entry point.
Update agent/tools/ to re-export from shared/, update test mock paths.
All 42 existing tests pass.

Closes #1
