+++ media_assistant/config.example.yaml
@@ -1,0 +1,31 @@
+# Media Assistant Configuration
+# Copy to config.yaml and adjust as needed.
+
+audio:
+  sample_rate: 16000
+  frame_size: 512
+  mic_device: null  # null = default mic
+
+aec:
+  enabled: true
+  filter_length: 1024
+  auto_mute_factor: 0.1  # reduce volume to 10% on wake word
+
+wake_word:
+  model_path: media_assistant/wakeword/models/jarvis.onnx
+  threshold: 0.8
+  energy_ratio_threshold: 1.5
+
+stt:
+  whisper_model: large-v3-turbo
+  whisper_device: cuda
+  whisper_compute_type: int8
+  vosk_model_path: models/vosk-model-small-ru-0.22
+  max_listen_seconds: 5.0
+
+llm_fallback:
+  enabled: true
+  ollama_url: http://localhost:11434
+  model: qwen3:4b
+
+browser_cdp_url: http://localhost:9222
